app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-firestore==2.14.0
google-cloud-storage==2.10.0
google-generativeai==0.3.0
google-cloud-aiplatform==1.39.0
python-dotenv==1.0.0
Pillow==10.1.0
aiofiles==23.2.1
cachetools==5.3.2
